[박정제] step-6 댓글

+++ src/main/java/cafe/questions/servlet/QuestionServlet.java
@@ -2,9 +2,10 @@ package cafe.questions.servlet;
 
 import cafe.MappingHttpServlet;
 import cafe.questions.Article;
+import cafe.questions.Reply;
 import cafe.questions.repository.ArticleRepository;
+import cafe.questions.repository.ReplyRepository;
 import cafe.users.User;
-import cafe.users.repository.UserRepository;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
@@ -19,16 +20,16 @@ import java.util.logging.Logger;
 public class QuestionServlet extends MappingHttpServlet {
     private static final Logger log = Logger.getLogger(QuestionServlet.class.getName());
     private final ArticleRepository articleRepository;
-    private final UserRepository userRepository;
+    private final ReplyRepository replyRepository;
 
     @Override
     public List<String> mappings() {
         return List.of("/questions/*");
     }
 
-    public QuestionServlet(ArticleRepository articleRepository, UserRepository userRepository) {
+    public QuestionServlet(ArticleRepository articleRepository, ReplyRepository replyRepository) {
         this.articleRepository = articleRepository;
-        this.userRepository = userRepository;
+        this.replyRepository = replyRepository;
     }
 
 
@@ -37,9 +38,8 @@ public class QuestionServlet extends MappingHttpServlet {
         Long id = Long.valueOf(req.getPathInfo().substring(1));
         Article article = articleRepository.findById(id);
         req.setAttribute("article", article);
-        if (article.getUserId() != null || article.getUserId() != 0) {
-            req.setAttribute("user", userRepository.findById(article.getUserId()));
-        }
+        List<Reply> replyList = replyRepository.findByArticleId(id);
+        req.setAttribute("replyList", replyList);
         req.getRequestDispatcher("/WEB-INF/views/questions/question.jsp").forward(req, resp);
     }
 
@@ -74,6 +74,16 @@ public class QuestionServlet extends MappingHttpServlet {
     protected void doDelete(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
         Long id = Long.valueOf(req.getPathInfo().substring(1));
         if (validAutorized(req, resp, articleRepository.findById(id))) return;
+
+        List<Reply> replyList = replyRepository.findByArticleId(id);
+        User user = (User) req.getSession().getAttribute("user");
+        if (!replyList.stream().allMatch(reply -> reply.getUserId().equals(user.getId()))) {
+            resp.setStatus(HttpServletResponse.SC_FORBIDDEN);
+            resp.getWriter().write("삭제할 수 없는 댓글이 있습니다.");
+            return;
+        }
+
+        replyList.forEach(reply -> replyRepository.deleteById(reply.getId()));
         articleRepository.deleteById(id);
     }
